Initial developments
The roots of this solution lie in the Estonian GDI node solution, however, this codebase has been extensively modified to make it standalone with custom dataset preparation tools. Now this solution is ready for early testing and feedback. The initial target version has been selected to be 0.9.0 to reflect its current maturity.